refactored structure of ui components
